PSG-1711: remove logging splunk handler as obsolete. Rename logging.py to logger.py to avoid circular dependency
