+++ pyproject.toml
@@ -14,12 +14,10 @@ pre-commit = "2.17.0"
 pytest = "7.1.3"
 pytest-socket = "^0.3.5"
 python = "^3.10"
-requests = "2.28.1"
 psutil = "5.9.0"
 pandas = "^1.4.2"
 marshmallow = "^3.15.0"
 structlog = "^21.5.0"
-types-requests = "^2.27.28"
 urllib3 = "1.26.12"
 GitPython = "^3.1.29"
