+++ scripts/synthetic/generate_results.py
@@ -9,7 +9,7 @@ from json import JSONEncoder
 import click
 import pandas as pd
 
-from scripts.util.logging import get_structlog_logger
+from scripts.util.logger import get_structlog_logger
 from scripts.util.metadata import EXPECTED_HEADERS as EXPECTED_METADATA_HEADERS, SAMPLE_ID, ILLUMINA, ONT
 from scripts.validation.check_csv_columns import check_csv_columns
