+++ tests/common/test_check_metadata.py
@@ -3,7 +3,7 @@ from click.testing import CliRunner
 import structlog
 
 from scripts.common.check_metadata import check_metadata, validate_metadata
-from scripts.util.logging import get_structlog_logger
+from scripts.util.logger import get_structlog_logger
 from tests.util.test_notification import load_log_file_to_dict
