+++ tests/util/test_logging.py
@@ -1,7 +1,7 @@
 import pytest
 import structlog
 
-from scripts.util.logging import LOG_LEVELS, get_structlog_logger
+from scripts.util.logger import LOG_LEVELS, get_structlog_logger
 from tests.util.test_notification import load_log_file_to_dict
 
 
@@ -33,15 +33,3 @@ def test_logger(tmp_path, message, level, sample):
     assert sample == log_sample["sample"]
     assert level == log_sample["level"]
     assert message == log_sample["message"]
-
-
-def test_logger_splunk_handler(tmp_path, setup_splunk):
-    message = "a message to be logged to Splunk"
-
-    log_file = tmp_path / "messages.log"
-    assert not log_file.is_file()
-
-    structlog.reset_defaults()
-    logger = get_structlog_logger(log_file=f"{log_file}")
-
-    logger.info(message)
