+++ tests/util/test_notification.py
@@ -5,7 +5,7 @@ import pytest
 import structlog
 
 from scripts.util.notifications import Event, Notification
-from scripts.util.logging import get_structlog_logger, ERROR, INFO
+from scripts.util.logger import get_structlog_logger, ERROR, INFO
 
 
 def load_log_file_to_dict(log_file: Path, key: str) -> Dict:
